# Conflicts:
#	src/main/java/com/example/final_project/Config/ConfigurationSecurity.java
#	src/main/java/com/example/final_project/Controller/ItemSaleController.java
#	src/main/java/com/example/final_project/Controller/SalesController.java
#	src/main/java/com/example/final_project/Service/ItemSaleService.java

+++ src/main/java/com/example/final_project/Service/ItemSaleService.java
@@ -26,6 +26,8 @@ public class ItemSaleService {
     private final ProductRepository productRepository;
     private final SalesRepository salesRepository;
     private final AccountantRepository accountantRepository;
+    private final SalesService salesService;
+
 
     public List<ItemSale> getAllItemSales() {
         return itemSaleRepository.findAll();
@@ -45,13 +47,35 @@ public class ItemSaleService {
         itemSaleRepository.save(itemSale);
     }
 
+    public void removeItemFromSale(Integer accounterId, Integer itemId,Integer saleId) {
 
-    public void deleteItemSale(Integer id) {
-        ItemSale itemSale = itemSaleRepository.findItemSaleById(id);
-                if(itemSale==null){
-                  throw new ApiException("ItemSale not found");
-                }
+        Accountant accountant = accountantRepository.findAccountantById(accounterId);
+        if (accountant == null) {
+            throw new ApiException("Accountant not found");
+        }
+        ItemSale itemSale = itemSaleRepository.findItemSaleById(itemId);
+        if (itemSale == null) {
+            throw new ApiException("ItemSale not found");
+        }
+        Sales currentSale = salesRepository.findSalesById(saleId);
+        if (currentSale == null) {
+            throw new ApiException("Sale not found");
+        }
         itemSaleRepository.delete(itemSale);
+
+        Product product1 = itemSale.getProduct();
+
+        if (currentSale.getSalesStatus().equals("CONFIRMED")) {
+            throw new ApiException("Cannot remove item from a confirmed invoice");
+        }
+
+
+        product1.setStock(product1.getStock()+itemSale.getQuantity());
+        salesService.updateCalculations(saleId);
+        productRepository.save(product1);
+        itemSaleRepository.delete(itemSale);
+        salesRepository.save(currentSale);
+
     }
 
     public List<ItemSale> getItemSalesBySalesId(Integer accountantId,Integer salesId) {
